chore: address reviewer feedback.
remove generator.
