+++ optimum/quanto/models/__init__.py
@@ -17,8 +17,6 @@ import os
 from collections.abc import Mapping
 from typing import Any, Dict, List, Optional, Union
 
-from safetensors import safe_open
-
 
 def is_transformers_available() -> bool:
     return importlib.util.find_spec("transformers") is not None
@@ -31,37 +29,3 @@ def is_diffusers_available() -> bool:
 if is_transformers_available():
     from .diffusers_models import *
     from .transformers_models import *
-
-
-class ShardedStateDict(Mapping):
-    """A pytorch state_dict stored in multiple safetensors files
-
-    This class implements the `collections.abc.Mapping` interface.
-    It can be passed to `torch.nn.Module.load_state_dict()` to recursively
-    load the module tensors.
-    """
-
-    def __init__(self, base_dir: str, tensor_index: Dict[str, str]):
-        self._base_dir = base_dir
-        self._index = tensor_index
-        self._handles = {}
-
-    def __iter__(self):
-        yield from self._index
-
-    def __len__(self):
-        return self._index.__len__()
-
-    def __getitem__(self, key: Any) -> Any:
-        filename = self._index.__getitem__(key)
-        if filename not in self._handles:
-            f = safe_open(os.path.join(self._base_dir, filename), framework="pytorch")
-            self._handles[filename] = f
-        f = self._handles[filename]
-        return f.get_tensor(key)
-
-    def __contains__(self, key: object) -> bool:
-        return self._index.__contains__(key)
-
-    def keys(self):
-        return self._index.keys()
