+++ optimum/quanto/models/diffusers_models.py
@@ -36,6 +36,8 @@ from diffusers.utils import (
     is_accelerate_available,
 )
 
+from .shared_dict import ShardedStateDict
+
 
 class QuantizedDiffusersModel:
 
@@ -149,8 +151,6 @@ class QuantizedDiffusersModel:
             # Look for the index of a sharded checkpoint
             checkpoint_file = os.path.join(model_name_or_path, SAFE_WEIGHTS_INDEX_NAME)
             if os.path.exists(checkpoint_file):
-                from . import ShardedStateDict
-
                 # Convert the checkpoint path to a list of shards
                 _, sharded_metadata = _get_checkpoint_shard_files(model_name_or_path, checkpoint_file)
                 # Create a mapping for the sharded safetensor files
